+++ Assets/Scripts/Enemy/EnemyDataSO/EnemyData.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "EnemyData", menuName = "ScriptableObjects/EnemyData", order = 2)]
 public class EnemyData : ScriptableObject
 {
-     public int Health = 1;
+     public float Health = 1;
      public int MaxHealth = 10;
 }
